generated-agents/web-research-agent/requirements.txt
zhipuai>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
ddgs>=0.1.0
